rename(account): constants 폴더명 변경
